Consolidate validation code into a single location
