+++ .changeset/spicy-pillows-sparkle.md
@@ -1,0 +1,5 @@
+---
+'@keystone-next/keystone': patch
+---
+
+Refactored mutation validation handling into a single location.
